feat(generator): allow {} as alias for `Record<string, never>`
Technically, the {} type means "value must be non-nullish", but in the case of a route handler’s data argument, this is always the case (the argument is never null or undefined). Therefore, since no practical use exists for the {} type in this case, we can treat it as a shorthand for `Record<string, never>`. Of course, some teams may choose not to use this "feature", if they think it’s too confusing or misleading.